Fixed loading images and normalizing 0 to 1

+++ loader_splitter/Loader.py
@@ -23,7 +23,7 @@ class Loader:
         Loads the images with their labels
         """
         n_images = sum([len(listdir(path.join(self._base_folder, category))) for category in self._categories])
-        samples = np.empty((n_images, self._image_size[0], self._image_size[1], 3), dtype=np.uint8)
+        samples = np.empty((n_images, self._image_size[0], self._image_size[1], 3), dtype=np.float16)
         labels = []
         image_index = 0
         for i, category in enumerate(self._categories):
@@ -31,8 +31,8 @@ class Loader:
             filenames = listdir(category_path)
             for j, filename in enumerate(filenames):
                 image_path = path.join(category_path, filename)
-                samples[image_index] = resize(imread(image_path).astype(np.uint8), self._image_size,
-                                              anti_aliasing=True)
+                samples[image_index] = resize(imread(image_path), self._image_size,
+                                              anti_aliasing=True).astype(np.float16)
                 labels.append(self._categories_encoded[i])
                 image_index += 1
         return SupervisedData(samples, np.array(labels))
